feat[#55]: 메인 페이지 보완
