+++ src/components/Main/MainPage.js
@@ -1,34 +1,14 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { fetchShowList } from '../../features/show/slices/showSlice';
-import './MainPage.css';
-import slides from './image.json';
-import getShows from '../../features/show/api/getShows';
-import Slider from './Slider';
+import React from 'react';
+import MainPageDesktop from './desktop/MainPageDesktop';
+import Desktop from '../MediaQuery/Desktop';
+// import Mobile from '../MediaQuery/Mobile';
+// import MainPageMobile from './mobile/MainPageMobile';
 
 function MainPage() {
-    const dispatch = useDispatch();
-    const showList = useSelector((state) => state.show.showList);
-    async function onRefresh() {
-        await getShows().then((value) => {
-            dispatch(fetchShowList(value));
-            console.log(showList[0]);
-        });
-    }
-
-    useEffect(() => {
-        onRefresh();
-    }, []);
-
     return (
-        <div>
-            <h3 className="ticketsOpen">📽️Tickets Open🎞</h3>
-            <div className="slider">
-                <div>
-                    <Slider slides={slides} />
-                </div>
-            </div>
-        </div>
+        <Desktop>
+            <MainPageDesktop />
+        </Desktop>
     );
 }
