+++ src/components/Main/desktop/MainPageDesktop.css
@@ -13,5 +13,5 @@
   border-style: groove;
   border-width: 6px;
   border-color:  rgba(43, 50, 82, 0.55);
-  background: rgba(43, 50, 82, 0.25) ;
+  background: rgba(43, 50, 82, 0.55);
 }
