+++ src/components/Main/image.json
@@ -11,7 +11,7 @@
   },
   {
     "image": "images/Dongari3.jpg",
-    "title": "메두사의 똇목",
+    "title": "메두사의 뗏목",
     "date": "2023.06.18(일) 22:00"
   },
   {
@@ -31,7 +31,7 @@
   },
   {
     "image": "images/Dongari3.jpg",
-    "title": "메두사의 똇목",
+    "title": "메두사의 뗏목",
     "date": "2023.06.18(일) 22:00"
   },
   {
